backend/requirements.txt
fastapi
uvicorn
pymongo
dnspython
certifi
python-dotenv
passlib[bcrypt]
bcrypt
python-jose[cryptography]
pandas
numpy
scikit-learn
tensorflow
passlib
bcrypt
python-jose
